feature/#59: 이웃 승인 로직 추가 및 파일 이름 변경
