--- src/neighbor/dto/update-neighbor.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/swagger';
-import { CreateNeighborDto } from './create-neighbor.dto';
-
-export class UpdateNeighborDto extends PartialType(CreateNeighborDto) {}
