+++ src/neighbors/dtos/update-neighbor.dto.ts
@@ -1,0 +1,4 @@
+export class UpdateNeighborDto {
+  no: number;
+  status: boolean;
+}
